+++ src/controllers/DefaultController.php
@@ -178,11 +178,9 @@ class DefaultController extends Controller
             $statusCode = $e instanceof HttpException ? $e->statusCode : 500;
             $statusText = $e->getMessage();
 
-            // log error
-            Craft::error($e->getMessage(), __METHOD__);
-            /** @var \yii\base\ErrorHandler $errorHandler */
-            $errorHandler = $this->getErrorHandler();
-            $errorHandler->logException($e);
+            // Log the exception
+            Craft::error('Error resolving Element API endpoint: ' . $e->getMessage(), __METHOD__);
+            Craft::$app->getErrorHandler()->logException($e);
         }
 
         // Create a JSON response formatter with custom options