+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Element API now requires Craft CMS 3.5 or later.
 - Endpoint responses are now cached by default, and are invalidated automatically when relevant elements are saved or deleted.
 
+### Fixed
+- Fixed a bug where non-200 responses were getting cached. ([#130](https://github.com/craftcms/element-api/issues/130))
+
 ## 2.6.0 - 2019-08-01
 
 ### Added